+++ src/NServiceBus.Metrics/ProbeBuilders/SignalProbeBuilder.cs
